requirements.txt
blinker==1.4
email-validator==1.1.1
flask==1.1.2
flask-bcrypt==0.7.1
flask-login==0.5.0
flask-migrate==2.5.3
flask-sqlalchemy==2.4.3
flask-wtf==0.14.3
gunicorn==20.0.4
psycopg2-binary==2.8.5
python-dotenv==0.13.0
sentry-sdk==0.14.4